+++ ui/UIButton.h
@@ -37,7 +37,7 @@ namespace Indigo
 			UI9Patch m_background;
 			glm::vec4 m_colour;
 			glm::vec4 m_shadow;
-			glm::u16vec2 m_drop;
+			glm::uvec2 m_drop;
 		};
 
 		template <typename Allocator>
